// packages/grpc/examples/probe_pb.ts
// @generated by protoc-gen-es v2.16.0 with parameter "target=ts"
// @generated from file probe.proto (package structure.probe, syntax proto3)
/* eslint-disable */

import type { Message } from "@bufbuild/protobuf";
import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv2";

/**
 * Describes the file probe.proto.
 */
export const file_probe: GenFile =
  /*@__PURE__*/
  fileDesc(
    "Cgtwcm9iZS5wcm90bxIPc3RydWN0dXJlLnByb2JlIhUKBFRleHQSDQoFdmFsdWUYASABKAky6wEKBVByb2JlEjUKBVVuYXJ5EhUuc3RydWN0dXJlLnByb2JlLlRleHQaFS5zdHJ1Y3R1cmUucHJvYmUuVGV4dBI3CgVXYXRjaBIVLnN0cnVjdHVyZS5wcm9iZS5UZXh0GhUuc3RydWN0dXJlLnByb2JlLlRleHQwARI4CgZVcGxvYWQSFS5zdHJ1Y3R1cmUucHJvYmUuVGV4dBoVLnN0cnVjdHVyZS5wcm9iZS5UZXh0KAESOAoEQ2hhdBIVLnN0cnVjdHVyZS5wcm9iZS5UZXh0GhUuc3RydWN0dXJlLnByb2JlLlRleHQoATABYgZwcm90bzM",
  );

/**
 * @generated from message structure.probe.Text
 */
export type Text = Message<"structure.probe.Text"> & {
  /**
   * @generated from field: string value = 1;
   */
  value: string;
};

/**
 * Describes the message structure.probe.Text.
 * Use `create(TextSchema)` to create a new message.
 */
export const TextSchema: GenMessage<Text> = /*@__PURE__*/ messageDesc(file_probe, 0);

/**
 * @generated from service structure.probe.Probe
 */
export const Probe: GenService<{
  /**
   * @generated from rpc structure.probe.Probe.Unary
   */
  unary: {
    methodKind: "unary";
    input: typeof TextSchema;
    output: typeof TextSchema;
  };
  /**
   * @generated from rpc structure.probe.Probe.Watch
   */
  watch: {
    methodKind: "server_streaming";
    input: typeof TextSchema;
    output: typeof TextSchema;
  };
  /**
   * @generated from rpc structure.probe.Probe.Upload
   */
  upload: {
    methodKind: "client_streaming";
    input: typeof TextSchema;
    output: typeof TextSchema;
  };
  /**
   * @generated from rpc structure.probe.Probe.Chat
   */
  chat: {
    methodKind: "bidi_streaming";
    input: typeof TextSchema;
    output: typeof TextSchema;
  };
}> = /*@__PURE__*/ serviceDesc(file_probe, 0);